[netty4] Extract apifest-api and fixes when read request content
