+++ apifest-api/src/main/java/com/apifest/api/AuthenticationException.java
