+++ apifest-api/src/main/java/com/apifest/api/BasicFilter.java
