+++ apifest-api/src/main/java/com/apifest/api/ExceptionEventHandler.java
