+++ apifest-api/src/main/java/com/apifest/api/GlobalErrors.java
