+++ apifest-api/src/main/java/com/apifest/api/IUserAuthentication.java
