+++ apifest-api/src/main/java/com/apifest/api/LifecycleHandler.java
