+++ apifest-api/src/main/java/com/apifest/api/Mapping.java
