+++ apifest-api/src/main/java/com/apifest/api/MappingAction.java
