+++ apifest-api/src/main/java/com/apifest/api/MappingEndpoint.java
