+++ apifest-api/src/main/java/com/apifest/api/MappingError.java
