+++ apifest-api/src/main/java/com/apifest/api/MappingException.java
