+++ apifest-api/src/main/java/com/apifest/api/UserDetails.java
