+++ apifest/pom.xml
@@ -69,6 +69,11 @@
   </build>
   <dependencies>
     <dependency>
+      <groupId>com.apifest</groupId>
+      <artifactId>apifest-api</artifactId>
+      <version>1.0.0-SNAPSHOT</version>
+    </dependency>
+    <dependency>
       <groupId>io.netty</groupId>
       <artifactId>netty-all</artifactId>
       <version>4.1.25.Final</version>
